fix(dsr,incident): harden DSR time validation and widen incident PK; un-skip guards
BUG DSR-1/2 (malformed time -> HTTP 500):
StoreDailyReportRequest now validates call_time and safety_meeting_time as
'nullable|date_format:H:i' (was bare 'nullable'). The DSR create form posts
HH:MM via <input type="time"> and both columns are MySQL `time` under STRICT,
so a malformed value (e.g. "notatime"/"xx:yy") previously bypassed validation
and threw 1292 at INSERT. It is now a validation error instead of a 500.
Chose H:i (not H:i:s) because the form emits HH:MM and the show view formats
the stored value with ->format('H:i').

BUG INC-02 (tinyint PK caps table at 255 rows):
Widened `id` from `tinyint(3) unsigned` to `int(10) unsigned` in both
create migrations:
  - 2026_06_25_000022_create_hazardnotifications_table.php
  - 2026_06_25_000045_create_unsafeconds_table.php
No FK widening needed: no child table references these tables. The only
related columns (related_unsafecond_id in hazardnotifications,
related_hazard_id in unsafeconds) already live in those same tables as
bigint with no FK constraint, so migrate:fresh has no type mismatch.

Regression guards un-skipped and converted to active assertions:
  - DsrBugGuardsTest: both time guards now assert not-500 + session errors
    for a malformed time, and acceptance of a valid "07:30".
  - IncidentPersistenceSealTest: INC-02 guard now asserts the id column type
    of both tables is no longer tinyint.

Verified against a throwaway crewcare_test_fix DB: 76 passed, 0 failed,
0 skipped/incomplete; migrate:fresh succeeded.

+++ database/migrations/2026_06_25_000022_create_hazardnotifications_table.php
@@ -10,7 +10,7 @@ class CreateHazardnotificationsTable extends Migration
     {
         DB::statement(<<<'SQL'
 CREATE TABLE `hazardnotifications` (
-  `id` tinyint(3) unsigned NOT NULL AUTO_INCREMENT,
+  `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
   `production_name` varchar(255) DEFAULT NULL,
   `name_loc` varchar(255) DEFAULT NULL,
   `latitude` decimal(10,7) DEFAULT NULL,
